requirements.txt
click==8.1.3
psycopg2-binary==2.9.3
ujson>=5.4.0
Unidecode==1.1.1
requests==2.28.1
sentry-sdk[flask]==0.19.3
typesense==0.14.0
python-Levenshtein==0.21.0
datasethoster@git+https://github.com/mayhem/data-set-hoster.git@v-2023-05-19.0
pysolr==3.9.0
ago==0.0.95
troi@git+https://github.com/metabrainz/troi-recommendation-playground@lb-radio
wheel
